requirements.txt
fastapi==0.95.0
uvicorn[standard]==0.22.0
faiss-cpu==1.7.3
transformers==4.31.0
datasets==2.14.0
boto3==1.28.0
numpy==1.25.0
pydantic==1.10.7
pytest==7.3.1
prometheus-fastapi-instrumentator==5.9.1
sentence-transformers==2.2.2
